requirements.txt

Flask==3.0.0
Werkzeug==3.1.3
Flask-Cors==3.0.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
click==8.1.8
blinker==1.9.0


PyJWT==2.10.1
python-dotenv==1.0.1
cryptography==42.0.5


redis==5.0.3
flask-limiter==3.5.1

SQLAlchemy==2.0.40
psycopg2-binary==2.9.10
greenlet==3.2.1

packaging==25.0
typing_extensions==4.13.2
six==1.17.0

fpdf==1.7.2

pytest==8.1.1
pytest-cov==5.0.0
iniconfig==2.1.0
coverage==7.8.0

gunicorn==21.2.0

setuptools==80.3.1
wheel==0.45.1
